5_results/requirements.txt
pydantic
tqdm
python-dotenv
ruff
matplotlib
seaborn
numpy
scipy
pandas
